refactor: prepare for multiple BVH construction algorithm support
