+++ clovers/src/objects/moving_sphere.rs
@@ -13,7 +13,7 @@ use rand::rngs::SmallRng;
 
 #[derive(Clone, Debug)]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
-/// `SphereInit` structure describes the necessary data for constructing a [`Sphere`](super::Sphere). Used with [serde] when importing [`SceneFile`](crate::scenes::SceneFile)s.
+/// `SphereInit` structure describes the necessary data for constructing a [`Sphere`](super::Sphere).
 pub struct MovingSphereInit {
     /// Used for multiple importance sampling
     #[cfg_attr(feature = "serde-derive", serde(default))]
