+++ clovers/src/objects/sphere.rs
@@ -14,7 +14,7 @@ use rand::{rngs::SmallRng, Rng};
 
 #[derive(Clone, Debug)]
 #[cfg_attr(feature = "serde-derive", derive(serde::Serialize, serde::Deserialize))]
-/// `SphereInit` structure describes the necessary data for constructing a [Sphere]. Used with [serde] when importing [`SceneFile`](crate::scenes::SceneFile)s.
+/// `SphereInit` structure describes the necessary data for constructing a [Sphere].
 pub struct SphereInit {
     /// Used for multiple importance sampling
     #[cfg_attr(feature = "serde-derive", serde(default))]
